refactor(order): 주문 페이지의 컴포넌트들을 분리시킴
분리시키고 작동하긴하는데 Order페이지가 한덩어리일때 진짜 손도 댈수없어서 임시로 분리시켰습니다. 로직 개선과 사용되는 state와 props들의 선언 위치를 변경하고 modal컴포넌트도 한번더 분리해야할 것같습니다
